+++ src/CustomizeBuildings/Changelog.md
@@ -5,6 +5,9 @@
 ### Added
 - pressured valve option that adds a button to limit the output pressure; very niche application
 
+### Fixed
+- fixed non-DLC config being in the wrong folder
+
 ## [1.0.27.0] EX1 S6-448916
 
 ### Fixed