+++ src/CustomizeGeyser/Changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.0.6.0]
+
+### Fixed
+- added missing slush_salt_water to the default list
+- fixed geysers being random even after RandomizerEnabled set to false; note that since the DLC a lot more geysers are predetermined, so randomizing only fully works on word generation!
+
 ## [1.0.5.0] EX1 S6-448916
 
 ### Fixed