+++ src/CustomizePlants/Changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.0.16.0] EX1 S6-449549
+
+### fixed
+- fixed safe_elements not being applied
+
 ## [1.0.15.0] EX1 S6-448916
 
 ### Added