+++ src/EggCritterSurplus/Changelog.md
@@ -1,6 +1,14 @@
 # Changelog
 
+## [1.0.6.0] EX1 S6-449549
+
+### Fixed
+- fixed Pacu not being able to be delivered
+
 ## [1.0.6.0] EX1 S6-448916
+
+### Changed
+- fetch jobs copy the buildings priority
 
 ### Fixed
 - updated for DLC1